+++ src/machinable/collection/__init__.py
@@ -1,0 +1,3 @@
+from machinable.collection.collection import Collection, collect
+from machinable.collection.record import RecordCollection
+from machinable.collection.submission import SubmissionCollection
