+++ src/machinable/submission/views/views.py
@@ -3,81 +3,74 @@ import inspect
 from machinable.utils.importing import ModuleClass
 from machinable.utils.utils import is_valid_variable_name
 
-_register = {
-    "submission": {},
-    "component": {},
-}
+_register = {}
 
 _used_attributes = {
-    "submission": {
-        "host",
-        "model",
-        "unique_id",
-        "create",
-        "timestamp",
-        "code_diff",
-        "code_backup",
-        "serialize",
-        "project_name",
-        "_cache",
-        "components",
-        "is_finished",
-        "submission_id",
-        "seed",
-        "submissions",
-        "find_many",
-        "started_at",
-        "code_version",
-        "file",
-        "finished_at",
-        "_model",
-        "experiment_name",
-        "ancestor",
-        "view",
-        "url",
-        "find",
-        "is_started",
-        "engine",
-    },
-    "component": {
-        "config",
-        "started_at",
-        "data",
-        "is_started",
-        "file",
-        "status",
-        "view",
-        "output",
-        "serialize",
-        "flags",
-        "get_records",
-        "records",
-        "is_finished",
-        "is_incomplete",
-        "state",
-        "has_records",
-        "_model",
-        "finished_at",
-        "create",
-        "heartbeat_at",
-        "components",
-        "submission",
-        "url",
-        "_cache",
-        "log",
-        "component_id",
-        "unique_id",
-        "is_active",
-        "tuning",
-        "host",
-        "engine",
-    },
+    "host",
+    "model",
+    "unique_id",
+    "create",
+    "timestamp",
+    "code_diff",
+    "code_backup",
+    "serialize",
+    "project_name",
+    "_cache",
+    "components",
+    "is_finished",
+    "submission_id",
+    "seed",
+    "submissions",
+    "find_many",
+    "started_at",
+    "code_version",
+    "file",
+    "finished_at",
+    "_model",
+    "experiment_name",
+    "ancestor",
+    "view",
+    "url",
+    "find",
+    "is_started",
+    "engine",
+    "config",
+    "started_at",
+    "data",
+    "is_started",
+    "file",
+    "status",
+    "view",
+    "output",
+    "serialize",
+    "flags",
+    "get_records",
+    "records",
+    "is_finished",
+    "is_incomplete",
+    "state",
+    "has_records",
+    "_model",
+    "finished_at",
+    "create",
+    "heartbeat_at",
+    "components",
+    "submission",
+    "url",
+    "_cache",
+    "log",
+    "component_id",
+    "unique_id",
+    "is_active",
+    "tuning",
+    "host",
+    "engine",
 }
 
 
-def get(view_type, instance, name=None):
+def get(instance, name=None):
     try:
-        return _register[view_type][name](instance)
+        return _register[name](instance)
     except KeyError:
         return None
