+++ tests/collections/collections_test.py
@@ -6,12 +6,7 @@ from unittest import TestCase
 
 import pytest
 from machinable import Submission
-from machinable.submission.collections import (
-    Collection,
-    SubmissionCollection,
-    SubmissionComponentCollection,
-    collect,
-)
+from machinable.collection import Collection, SubmissionCollection, collect
 
 
 def test_collect():
